[REF] Tests: move List tests to List plugin
