+++ packages/core/test/Parser.test.ts
@@ -8,8 +8,6 @@ import { LineBreakNode } from '../../plugin-linebreak/LineBreakNode';
 import { Heading } from '../../plugin-heading/Heading';
 import { Paragraph } from '../../plugin-paragraph/Paragraph';
 import { VElement } from '../src/VNodes/VElement';
-import { List } from '../../plugin-list/List';
-import { ListNode, ListType } from '../../plugin-list/ListNode';
 
 describe('utils', () => {
     describe('Parser', () => {
@@ -19,7 +17,6 @@ describe('utils', () => {
             ...LineBreak.parsingFunctions,
             ...Heading.parsingFunctions,
             ...Paragraph.parsingFunctions,
-            ...List.parsingFunctions,
         );
         describe('parse()', () => {
             it('should parse a "p" tag with some content', () => {
@@ -91,138 +88,6 @@ describe('utils', () => {
                     bold: false,
                     italic: true,
                     underline: false,
-                });
-            });
-            describe('Lists', () => {
-                it('should parse a complex list', () => {
-                    const element = document.createElement('div');
-                    element.innerHTML = [
-                        '<ul>',
-                        '    <li>', // li0: becomes P
-                        '        a',
-                        '        <ul>', // li1
-                        '            <li>         ', // li1_0: becomes P
-                        '                <b>a.</b>a',
-                        '            </li>',
-                        '            <li><p>a.b</p></li>', // li1_1
-                        '            <li><h1>a.c</h1></li>', // li1_2
-                        '            <li>a.d</li>', // li1_3: becomes P
-                        '            <li>',
-                        '                <ul>', // li1_4
-                        '                    <li>a.d.a</li>', // li1_4_0: becomes P
-                        '                </ul>',
-                        '            </li>',
-                        '        </ul>',
-                        '    </li>',
-                        '    <li>b</li>', // li2: becomes P
-                        '    <ol>', // li3
-                        '        <li>b.1</li>', // li3_0: becomes P
-                        '        <li><p>b.2</p></li>', // li3_1
-                        '        <li><h1>b.3</h1></li>', // li3_2
-                        '        <li>b.4</li>', // li3_3
-                        '    </ol>',
-                        '</ul>',
-                    ].join('\n');
-                    const vDocument = parser.parse(element);
-
-                    expect(vDocument.root.children.length).to.equal(1);
-                    const list = vDocument.root.firstChild() as ListNode;
-                    expect(list.toString()).to.equal('ListNode: unordered');
-                    expect(list.listType).to.equal(ListType.UNORDERED);
-                    expect(list.children.length).to.equal(4);
-
-                    const li0 = list.nthChild(0);
-                    expect(li0.toString()).to.equal('ParagraphNode');
-                    expect(li0.children.length).to.equal(1);
-                    expect(li0.firstChild().toString()).to.equal('a');
-
-                    const li1 = list.nthChild(1) as ListNode;
-                    expect(li1.toString()).to.equal('ListNode: unordered');
-                    expect(li1.listType).to.equal(ListType.UNORDERED);
-                    expect(li1.children.length).to.equal(5);
-
-                    /* eslint-disable @typescript-eslint/camelcase */
-                    const li1_0 = li1.nthChild(0);
-                    expect(li1_0.toString()).to.equal('ParagraphNode');
-                    expect(li1_0.children.length).to.equal(3);
-                    expect(li1_0.nthChild(0).toString()).to.equal('a');
-                    expect(li1_0.nthChild(1).toString()).to.equal('.');
-                    expect((li1_0.nthChild(1) as CharNode).format.bold).to.be.true;
-                    expect(li1_0.nthChild(2).toString()).to.equal('a');
-
-                    const li1_1 = li1.nthChild(1);
-                    expect(li1_1.toString()).to.equal('ParagraphNode');
-                    expect(li1_1.children.length).to.equal(3);
-                    expect(li1_1.nthChild(0).toString()).to.equal('a');
-                    expect(li1_1.nthChild(1).toString()).to.equal('.');
-                    expect(li1_1.nthChild(2).toString()).to.equal('b');
-
-                    const li1_2 = li1.nthChild(2);
-                    expect(li1_2.toString()).to.equal('HeadingNode: 1');
-                    expect(li1_2.children.length).to.equal(3);
-                    expect(li1_2.nthChild(0).toString()).to.equal('a');
-                    expect(li1_2.nthChild(1).toString()).to.equal('.');
-                    expect(li1_2.nthChild(2).toString()).to.equal('c');
-
-                    const li1_3 = li1.nthChild(3);
-                    expect(li1_3.toString()).to.equal('ParagraphNode');
-                    expect(li1_3.children.length).to.equal(3);
-                    expect(li1_3.nthChild(0).toString()).to.equal('a');
-                    expect(li1_3.nthChild(1).toString()).to.equal('.');
-                    expect(li1_3.nthChild(2).toString()).to.equal('d');
-
-                    const li1_4 = li1.nthChild(4) as ListNode;
-                    expect(li1_4.toString()).to.equal('ListNode: unordered');
-                    expect(li1_4.listType).to.equal(ListType.UNORDERED);
-                    expect(li1_4.children.length).to.equal(1);
-
-                    const li1_4_0 = li1_4.firstChild();
-                    expect(li1_4_0.toString()).to.equal('ParagraphNode');
-                    expect(li1_4_0.children.length).to.equal(5);
-                    expect(li1_4_0.nthChild(0).toString()).to.equal('a');
-                    expect(li1_4_0.nthChild(1).toString()).to.equal('.');
-                    expect(li1_4_0.nthChild(2).toString()).to.equal('d');
-                    expect(li1_4_0.nthChild(3).toString()).to.equal('.');
-                    expect(li1_4_0.nthChild(4).toString()).to.equal('a');
-
-                    const li2 = list.nthChild(2);
-                    expect(li2.toString()).to.equal('ParagraphNode');
-                    expect(li2.children.length).to.equal(1);
-                    expect(li2.firstChild().toString()).to.equal('b');
-
-                    const li3 = list.nthChild(3) as ListNode;
-                    expect(li3.toString()).to.equal('ListNode: ordered');
-                    expect(li3.listType).to.equal(ListType.ORDERED);
-                    expect(li3.children.length).to.equal(4);
-
-                    const li3_0 = li3.nthChild(0);
-                    expect(li3_0.toString()).to.equal('ParagraphNode');
-                    expect(li3_0.children.length).to.equal(3);
-                    expect(li3_0.nthChild(0).toString()).to.equal('b');
-                    expect(li3_0.nthChild(1).toString()).to.equal('.');
-                    expect(li3_0.nthChild(2).toString()).to.equal('1');
-
-                    const li3_1 = li3.nthChild(1);
-                    expect(li3_1.toString()).to.equal('ParagraphNode');
-                    expect(li3_1.children.length).to.equal(3);
-                    expect(li3_1.nthChild(0).toString()).to.equal('b');
-                    expect(li3_1.nthChild(1).toString()).to.equal('.');
-                    expect(li3_1.nthChild(2).toString()).to.equal('2');
-
-                    const li3_2 = li3.nthChild(2);
-                    expect(li3_2.toString()).to.equal('HeadingNode: 1');
-                    expect(li3_2.children.length).to.equal(3);
-                    expect(li3_2.nthChild(0).toString()).to.equal('b');
-                    expect(li3_2.nthChild(1).toString()).to.equal('.');
-                    expect(li3_2.nthChild(2).toString()).to.equal('3');
-
-                    const li3_3 = li3.nthChild(3);
-                    expect(li3_3.toString()).to.equal('ParagraphNode');
-                    expect(li3_3.children.length).to.equal(3);
-                    expect(li3_3.nthChild(0).toString()).to.equal('b');
-                    expect(li3_3.nthChild(1).toString()).to.equal('.');
-                    expect(li3_3.nthChild(2).toString()).to.equal('4');
-                    /* eslint-enable @typescript-eslint/camelcase */
                 });
             });
         });
